+++ CHANGES.md
@@ -10,6 +10,7 @@
       VPN profiles (for now shows warning if there is a problem)
     - make sure `range` is `/29` or lower (OpenVPN limitation)
     - make sure `range6` is `/112` or lower (OpenVPN limitation)
+- implement changes for updated `Config` API
 
 ## 2.2.4 (2020-09-08)
 - add `ECDSA` certificate support for TLSv1.2, already supported on TLSv1.3